+++ Doc/fake_ldap_module_for_documentation.py
@@ -1,21 +1,21 @@
 """
-A module that mocks `_ldap` for the purposes of generating documentation
+A module that mocks `ldap._ldap` for the purposes of generating documentation
 
-This module provides placeholders for the contents of `_ldap`, making it
-possible to generate documentation even _ldap is not compiled.
+This module provides placeholders for the contents of `ldap._ldap`, making it
+possible to generate documentation even if ldap._ldap is not compiled.
 It should also make the documentation independent of which features are
 available in the system OpenLDAP library.
 
 The overly long module name will show up in AttributeError messages,
-hinting that this is not the actual _ldap.
+hinting that this is not the actual ldap._ldap.
 
 See https://www.python-ldap.org/ for details.
 """
 
 import sys
 
-# Cause `import _ldap` to import this module instead of the actual `_ldap`.
-sys.modules['_ldap'] = sys.modules[__name__]
+# Cause `import ldap._ldap` to import this module instead of the actual module.
+sys.modules['ldap._ldap'] = sys.modules[__name__]
 
 from constants import CONSTANTS
 from pkginfo import __version__
