+++ Lib/ldap/__init__.py
@@ -34,10 +34,10 @@ else:
   _trace_file = sys.stderr
   _trace_stack_limit = None
 
-import _ldap
+import ldap._ldap as _ldap
 assert _ldap.__version__==__version__, \
        ImportError(f'ldap {__version__} and _ldap {_ldap.__version__} version mismatch!')
-from _ldap import *
+from ldap._ldap import *
 # call into libldap to initialize it right now
 LIBLDAP_API_INFO = _ldap.get_option(_ldap.OPT_API_INFO)
