+++ Lib/ldap/controls/__init__.py
@@ -10,7 +10,7 @@ Each class provides support for a certain control.
 
 from ldap.pkginfo import __version__
 
-import _ldap
+import ldap._ldap as _ldap
 assert _ldap.__version__==__version__, \
        ImportError(f'ldap {__version__} and _ldap {_ldap.__version__} version mismatch!')
