+++ MANIFEST.in
@@ -2,7 +2,7 @@ include MANIFEST.in Makefile CHANGES INSTALL LICENCE README TODO
 include tox.ini
 include Modules/*.c Modules/*.h
 include Lib/ldap/py.typed
-include Lib/_ldap.pyi
+include Lib/ldap/_ldap.pyi
 recursive-include Build *.cfg*
 recursive-include Lib *.py
 recursive-include Demo *.py
