+++ setup.cfg
@@ -20,7 +20,7 @@ defines = HAVE_SASL HAVE_TLS
 
 extra_compile_args =
 extra_objects =
-extra_files = .:Lib/_ldap.pyi
+extra_files = ./ldap/:Lib/ldap/_ldap.pyi
 
 # Uncomment this if your libldap is not thread-safe and you need libldap_r
 # instead
